+++ src/ONTraC/GNN/_GNN.py
@@ -16,9 +16,9 @@ from ..train import SubBatchTrainProtocol
 
 def load_data(options: Values) -> Tuple[SpatailOmicsDataset, DenseDataLoader]:
     """
-    Load data and create sample loader
-    :param options: options
-    :return: dataset, sample_loader
+    Load data and create sample loader.
+    :param options: options.
+    :return: dataset, sample_loader.
     """
 
     info('Loading dataset.')
@@ -32,8 +32,9 @@ def load_data(options: Values) -> Tuple[SpatailOmicsDataset, DenseDataLoader]:
 
 def set_seed(seed: int) -> None:
     """
-    Set seed
-    :param seed: seed
+    Set seed.
+    :param seed: seed.
+    :return: None.
     """
 
     random.seed(seed)
@@ -47,6 +48,16 @@ def train(options: Values,
           sample_loader: DenseDataLoader,
           inspect_funcs: Optional[List[Callable]] = None,
           model_name: str = 'GNN') -> SubBatchTrainProtocol:
+    """
+    GNN training process.
+    :param options: options.
+    :param nn_model: torch.nn.Module, GNN model.
+    :param BatchTrain: Type[SubBatchTrainProtocol], batch train class.
+    :param sample_loader: DenseDataLoader, sample loader.
+    :param inspect_funcs: list of inspect functions.
+    :param model_name: str, model name.
+    :return: batch_train.
+    """
     optimizer = torch.optim.Adam(nn_model.parameters(), lr=options.lr)
     batch_train = BatchTrain(model=nn_model, device=torch.device(options.device),
                              data_loader=sample_loader)  # type: ignore
@@ -73,7 +84,9 @@ def train(options: Values,
 def evaluate(batch_train: SubBatchTrainProtocol, model_name: str) -> None:
     """
     Evaluate the performance of ONTraC model on data.
-    :return: None
+    :param batch_train: SubBatchTrainProtocol, batch train.
+    :param model_name: str, model name.
+    :return: None.
     """
     info(message=f'Evaluating process start.')
     loss_dict: Dict[str, np.floating] = batch_train.evaluate()  # type: ignore
@@ -83,6 +96,14 @@ def evaluate(batch_train: SubBatchTrainProtocol, model_name: str) -> None:
 
 def predict(output_dir: str, batch_train: SubBatchTrainProtocol, dataset: SpatailOmicsDataset,
             model_name: str) -> Tuple[Optional[np.ndarray], Optional[np.ndarray]]:
+    """
+    Predict the results of ONTraC model on data.
+    :param output_dir: str, output directory.
+    :param batch_train: SubBatchTrainProtocol, batch train.
+    :param dataset: SpatailOmicsDataset, dataset.
+    :param model_name: str, model name.
+    :return: consolidate_s_array, consolidate_out_adj_array.
+    """
     info(f'Predicting process start.')
     each_sample_loader = DenseDataLoader(dataset, batch_size=1)
     consolidate_flag = False
@@ -141,9 +162,12 @@ def save_graph_pooling_results(ori_data_df: pd.DataFrame, dataset: SpatailOmicsD
                                consolidate_s_array: np.ndarray, output_dir: str) -> None:
     """
     Save graph pooling results as the Niche cluster (max probability for each niche & cell).
-    :param ori_data_df: pd.DataFrame, original data
-    :param consolidate_s_array: np.ndarray, consolidate s array
-    :return: None
+    :param ori_data_df: pd.DataFrame, original data. Sample and Cell_ID columns are used.
+    :param dataset: SpatailOmicsDataset, dataset.
+    :param rel_params: dict, relative parameters.
+    :param consolidate_s_array: np.ndarray, consolidate s array.
+    :param output_dir: str, output directory.
+    :return: None.
     """
 
     consolidate_s_niche_df = pd.DataFrame()