+++ src/ONTraC/optparser/_GNN.py
@@ -1,4 +1,5 @@
 from optparse import OptionParser, Values
+from tkinter import W
 
 from ..log import *
 from ..version import __version__
@@ -8,23 +9,23 @@ from ._train import *
 # ------------------------------------
 # Constants
 # ------------------------------------
-IO_OPTIONS = ['dataset', 'preprocessing_dir', 'GNN_dir', 'NTScore_dir']
+IO_OPTIONS = ['dataset', 'preprocessing_dir', 'GNN_dir']
 
 
 # ------------------------------------
 # Functions
 # ------------------------------------
-def prepare_GP_optparser() -> OptionParser:
+def prepare_GNN_optparser() -> OptionParser:
     """
     Prepare optparser object. New options will be added in thisfunction first.
     :return: OptionParser object.
     """
-    usage = f'''USAGE: %prog <-d DATASET> <--preprocessing-dir PREPROCESSING_DIR> <--GNN-dir GNN_DIR> <--NTScore-dir NTSCORE_DIR> 
-    [--device DEVICE] [--epochs EPOCHS] [--patience PATIENCE] [--min-delta MIN_DELTA] [--min-epochs MIN_EPOCHS] [--batch-size BATCH_SIZE] 
+    usage = f'''USAGE: %prog <--preprocessing-dir PREPROCESSING_DIR> <--GNN-dir GNN_DIR> [--device DEVICE]
+    [--epochs EPOCHS] [--patience PATIENCE] [--min-delta MIN_DELTA] [--min-epochs MIN_EPOCHS] [--batch-size BATCH_SIZE] 
     [-s SEED] [--seed SEED] [--lr LR] [--hidden-feats HIDDEN_FEATS] [-k K_CLUSTERS]
     [--modularity-loss-weight MODULARITY_LOSS_WEIGHT] [--purity-loss-weight PURITY_LOSS_WEIGHT] 
     [--regularization-loss-weight REGULARIZATION_LOSS_WEIGHT] [--beta BETA]'''
-    description = 'GP (Graph Pooling): GNN & Node Pooling'
+    description = 'Graph Neural Network (GNN)'
 
     # option processor
     optparser = OptionParser(version=f'%prog {__version__}',
@@ -43,7 +44,7 @@ def prepare_GP_optparser() -> OptionParser:
     return optparser
 
 
-def opt_GP_validate(optparser: OptionParser) -> Values:
+def opt_GNN_validate(optparser: OptionParser) -> Values:
     """Validate options from a OptParser object.
 
     :param optparser: OptionParser object.