+++ src/ONTraC/run/processes.py
@@ -5,6 +5,7 @@ import numpy as np
 import pandas as pd
 import torch
 
+from ..data import load_dataset
 from ..GNN import (evaluate, load_data, predict, save_graph_pooling_results,
                    set_seed, train)
 from ..log import *
@@ -17,10 +18,10 @@ from ..utils import get_rel_params, read_yaml_file
 
 def load_parameters(opt_validate_func: Callable, prepare_optparser_func: Callable) -> Values:
     """
-    Load parameters
-    :param opt_validate_func: validate function
-    :param prepare_optparser_func: prepare optparser function
-    :return: options
+    Load parameters.
+    :param opt_validate_func: validate function.
+    :param prepare_optparser_func: prepare optparser function.
+    :return: options.
     """
     options = opt_validate_func(prepare_optparser_func())
 
@@ -29,10 +30,10 @@ def load_parameters(opt_validate_func: Callable, prepare_optparser_func: Callabl
 
 def niche_network_construct(options: Values, ori_data_df: pd.DataFrame) -> None:
     """
-    Niche network construct process
-    :param options: options
-    :param ori_data_df: pd.DataFrame, original data
-    :return: None
+    Niche network construct process.
+    :param options: options.
+    :param ori_data_df: pd.DataFrame, original data.
+    :return: None.
     """
 
     info('------------- Niche network construct --------------- ')
@@ -47,10 +48,13 @@ def niche_network_construct(options: Values, ori_data_df: pd.DataFrame) -> None:
 def gnn(options: Values, ori_data_df: pd.DataFrame, nn_model: Type[torch.nn.Module],
         BatchTrain: Type[SubBatchTrainProtocol], inspect_funcs: Optional[List[Callable]]) -> None:
     """
-    GNN training and prediction process
-    :param options: options
-    :param ori_data_df: pd.DataFrame, original data
-    :return: None
+    GNN training and prediction process.
+    :param options: options.
+    :param ori_data_df: pd.DataFrame, original data.
+    :param nn_model: nn model.
+    :param BatchTrain: batch train.
+    :param inspect_funcs: inspect functions.
+    :return: None.
     """
 
     info('------------------------ GNN ------------------------ ')
@@ -88,11 +92,9 @@ def gnn(options: Values, ori_data_df: pd.DataFrame, nn_model: Type[torch.nn.Modu
 
 def NTScore(options: Values) -> None:
     """
-    Pseudotime calculateion process
-    :param options: options
-    :param consolidate_s_array: consolidate s array
-    :param consolidate_out_adj_array: consolidate out adj array
-    :return: None
+    Pseudotime calculateion process.
+    :param options: options.
+    :return: None.
     """
 
     info('----------------- Niche trajectory ------------------ ')
@@ -100,10 +102,11 @@ def NTScore(options: Values) -> None:
 
     params = read_yaml_file(f'{options.preprocessing_dir}/samples.yaml')
     rel_params = get_rel_params(options, params)
-    dataset, _ = load_data(options=options)
+    dataset = load_dataset(options=options)
 
     niche_cluster_score, niche_level_NTScore = get_niche_NTScore(niche_cluster_loading=consolidate_s_array,
-                                                                 niche_adj_matrix=consolidate_out_adj_array)
+                                                                 niche_adj_matrix=consolidate_out_adj_array,
+                                                                 options=options)
     cell_level_NTScore, all_niche_level_NTScore_dict, all_cell_level_NTScore_dict = niche_to_cell_NTScore(
         dataset=dataset, rel_params=rel_params, niche_level_NTScore=niche_level_NTScore)