+++ tests/test_niche_trajectory.py
@@ -4,14 +4,26 @@ from ONTraC.niche_trajectory._niche_trajectory import (
     get_niche_trajectory_path, trajectory_path_to_NC_score, get_niche_NTScore)
 
 
-def test_get_niche_trajectory_path():
+def test_get_niche_trajectory_path() -> None:
+    """
+    Test the function get_niche_trajectory_path.
+    
+    :return: None.
+    """
+
     # Test case: Adjacency matrix with 5 nodes
     adj_matrix = np.array([[0, 1, 0, 1, 0], [1, 0, 1, 0, 1], [0, 1, 0, 1, 0], [1, 0, 1, 0, 1], [0, 1, 0, 1, 0]])
     expected_path = [0, 1, 2, 3, 4]
     assert get_niche_trajectory_path(adj_matrix) == expected_path
 
 
-def test_trajectory_path_to_NC_score():
+def test_trajectory_path_to_NC_score() -> None:
+    """
+    Test the function trajectory_path_to_NC_score.
+
+    :return: None.
+    """
+
     # Test case: Niche trajectory path with 6 clusters
     niche_trajectory_path = [1, 2, 3, 4, 5, 0]
     expected_NT_score = np.array([1., 0., 0.2, 0.4, 0.6, 0.8])
@@ -19,7 +31,13 @@ def test_trajectory_path_to_NC_score():
     assert np.allclose(gen_NT_score, expected_NT_score)
 
 
-def test_get_niche_NTScore():
+def test_get_niche_NTScore() -> None:
+    """
+    Test the function get_niche_NTScore.
+
+    :return: None.
+    """
+    
     # Test case: Niche cluster loading with 3 clusters and adjacency matrix with 3 nodes
     niche_cluster_loading = np.array([[0.1, 0.2, 0.7], [0.4, 0.5, 0.1], [0.6, 0.1, 0.3]])  # #niche x #niche_cluster
     niche_adj_matrix = np.array([[0, 1, 0], [1, 0, 1], [0, 1, 0]])